requirements.txt
Flask
mongoengine
flask_mongoengine
flask_debugtoolbar
pymongo==3.5.1
python-docx
requests